add a workflow to automatically upload to pypi
This workflow can upload the package to pypi whenever a tagged release is pushed to master.  Besides, we can also trigger the workflow manually. 
We need to add a token to this repo before we can use it, as mentioned in https://packaging.python.org/guides/publishing-package-distribution-releases-using-github-actions-ci-cd-workflows/

The nightly release is not prepared yet. Because pypi has a total size limit for each project, I think maybe we don't need to upload it nightly. Actually, building the python part is very easy, users should be able to do it. The problem is to build the java & scala part, as they have many dependencies.  Therefore, maybe we can release jar separately, and tell users how to build the package with a built jar, or modify the path to jar.